+++ src/EntityGraphQL/Schema/GraphQLIgnoreAttribute.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace EntityGraphQL.Schema
+{
+    /// <summary>
+    /// Tell the Schema Builder to ignore this field or property
+    /// </summary>
+    public class GraphQLIgnoreAttribute : Attribute
+    {
+
+        public GraphQLIgnoreAttribute()
+        {
+        }
+    }
+}